\name{qpcRdefault}
\docType{data}
\alias{qpcRdefault}
\title{The processed data generated using the default version of the
  qpcR software.}
\description{These data are also used to demonstrate the functionality
  of the miRcomp package.}
\usage{data(qpcRdefault)}

\keyword{datasets}
\examples{
  data(qpcRdefault)
}
